+++ src/client/java/dev/mg95/xungeontimer/XungeonTimerClient.java
@@ -3,19 +3,26 @@ package dev.mg95.xungeontimer;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.command.v2.ClientCommandManager;
 import net.fabricmc.fabric.api.client.command.v2.ClientCommandRegistrationCallback;
+import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.message.v1.ClientReceiveMessageEvents;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayConnectionEvents;
 import net.fabricmc.fabric.api.client.rendering.v1.HudRenderCallback;
 import net.minecraft.client.MinecraftClient;
+import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.text.Text;
 import org.apache.commons.lang3.time.StopWatch;
+import dev.mg95.xungeontimer.TimerConfig;
 
 import java.util.Objects;
 
 public class XungeonTimerClient implements ClientModInitializer {
     private StopWatch timer = new StopWatch();
-
     private String username;
+    private boolean inXungeon = false;
+    private boolean wasDead = false;
+
+
+    public static final TimerConfig CONFIG = TimerConfig.createAndLoad();
 
     @Override
     public void onInitializeClient() {
@@ -32,6 +39,19 @@ public class XungeonTimerClient implements ClientModInitializer {
             } else if (Objects.equals(message.getString(), String.format("%s has escaped the Prison!", username))) {
                 stopTimer();
             }
+        });
+
+        ClientTickEvents.END_CLIENT_TICK.register(client -> {
+            if (!inXungeon || !CONFIG.resetOnDeath()) return;
+            PlayerEntity player = MinecraftClient.getInstance().player;
+            if (player == null) return;
+
+            if (player.isAlive() && wasDead) startTimer();
+            else if (player.isDead() && !timer.isStopped()) {
+                resetTimer();
+                inXungeon = true;
+            }
+            wasDead = player.isDead();
         });
 
 
@@ -53,11 +73,13 @@ public class XungeonTimerClient implements ClientModInitializer {
     }
 
     public void resetTimer() {
+        inXungeon = false;
         timer.reset();
     }
 
     public void startTimer() {
         resetTimer();
+        inXungeon = true;
         timer.start();
     }