+++ changelog.md
@@ -1,2 +1,4 @@
- # Xungeon Timer 1.0
- First release! 
+ # Xungeon Timer 1.1
+- Added config screen
+- Added the restart on death option
+- Added the timer position option